Bug with gunfire where sizeX/sizeY equals 2 and bug with unable to start new game after destroying playernetwork in the gamescene..

+++ Assets/Scripts/Network/PlayerNetwork.cs
@@ -75,7 +75,7 @@ public class PlayerNetwork : MonoBehaviour
     {
         int mask = 15;
 
-        shootingArea.sizeX = (area >> 4) & mask;
-        shootingArea.sizeY = area & mask;
+        shootingArea.sizeX = (float)((area >> 4) & mask);
+        shootingArea.sizeY = (float)(area & mask);
     }
 }
